% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_XYZ_functions.R
\name{ABC}
\alias{ABC}
\title{Function to perform ABC-analysis}
\usage{
ABC(mydata, ncol_what, ncol_by, a = 0.6, b = 0.2)
}
\arguments{
\item{mydata}{Data.frame with items sales}

\item{ncol_what}{Number of column where analysis should be performed}

\item{ncol_by}{number of column, that data should be aggregated by}

\item{a, b}{Shares coefficients of A and B groups}
}
\value{
list with sum value for \code{ncol_what} by values from \code{ncol_by}, divided to A, B, C groups
}
\description{
Function to perform ABC-analysis
}

