% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned.R
\name{binned}
\alias{binned}
\title{binned}
\usage{
binned(x, nbin = 0, type = "width")
}
\arguments{
\item{nbin}{the number of bins or an increasing vector giving the bin boundaries}

\item{type}{c("height", "width") either the same number of elements in a bin or equal widths for the bins}
}
\value{
the vector mapped to bin numbers 1 to nbin
}
\description{
map vector to bins
}
\examples{
# a<- binned(x, 10) # 10 equal width bins
# a<- binned(x, 0:10) # ten bins for numbers between 0-10
#                     # other numbers that do not fit are in the first or last bin
}

