% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance_modeling.R
\name{compute.irradiance.on.surface}
\alias{compute.irradiance.on.surface}
\title{compute.irradiance.on.surface}
\usage{
compute.irradiance.on.surface(
  potential.irradiance.J,
  surface.normals,
  surface.matrix,
  plot.constants,
  date.variables,
  time.interval.sec,
  raster.resolution
)
}
\arguments{
\item{plot.constants}{c(latitude,longitude,altitude,timezone)}

\item{date.variables}{c(julian.date,visibility.km,humidity,temperature.K,ozone.thickness,albedo)}
}
\description{
compute irradiance on real surface
}
\examples{
model.irradiance ()
}
\keyword{irradiance}
