% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz_trajectory2vector.R
\name{convertyXYZtrajectory2GEOJSON}
\alias{convertyXYZtrajectory2GEOJSON}
\title{convertyXYZtrajectory2GEOJSON}
\usage{
convertyXYZtrajectory2GEOJSON(input, output)
}
\arguments{
\item{string}{output}
}
\description{
Function to convert xyz trajectory file to geojson
}
\examples{
get_vuk_palette()
}
\keyword{trajectory}
