% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Do_Clustering_function.R
\name{Do.Clustering}
\alias{Do.Clustering}
\title{Function to do clustering of rows in data.frame#'}
\usage{
Do.Clustering(data, grid_x_dim = 10, grid_y_dim = 10, nclust = 10,
  show_plot = TRUE)
}
\arguments{
\item{data}{data.frame with numeric values, columns are features on which the clustering is based}

\item{grid_x_dim, grid_y_dim}{dimensions of the grid in somgrid}

\item{nclust}{amount of clusters needed}

\item{show_plot}{whether or not to show a clustering plot. Default is TRUE#'}
}
\value{
a vector comprising number of cluster for each data row.
}
\description{
Function to do clustering of rows in data.frame#'
}

