% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_likert.R
\name{fake_likert}
\alias{fake_likert}
\title{fake_likert}
\usage{
fake_likert(N = 16, M = 5, m = 3, d = 1, s = 2, rep.count = 250000,
  eps = 0.005, base = 0, T = 4, test = FALSE)
}
\arguments{
\item{N}{the number of observations}

\item{M}{the highest score}

\item{m}{the mean to approach}

\item{d}{the number of scores from base to M that are averaged (usually 1)}

\item{s}{the standard deviation to approach}

\item{rep.count}{the maximum number of try's}

\item{eps}{the tolerance for |m-s|}

\item{base}{the lowest value of the score (usually 0 or 1)}

\item{T}{the temperature (set to higher values below sqrt(rep.count) to find rare solutions)}

\item{test}{set to TRUE to print some test information from the fucntion}
}
\value{
a vector with N integers on the likert scale
}
\description{
Create a sample for a likert scale score with given mean and st.dev.
    This uses a primitive type of simulated annealing for the search. This was written
    to test the existence of fraudulent numbers in a manuscript by Diederik Stapel.
}
\examples{
# fake_likert(N=16, M=7, base=1, m=4.6, s=3.6) # Cannot exist. Should fail
}

