% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance_modeling.R
\name{get.julian.day.time}
\alias{get.julian.day.time}
\title{get.julian.day.time}
\usage{
get.julian.day.time(date)
}
\description{
convert date from "YYYY-MM-DDTHH-MM-SS" to julian
}
\examples{
get.julian.day.time()
}
\keyword{convert}
\keyword{date}
\keyword{julian}
