% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance_modeling.R
\name{load.meteodata.csv}
\alias{load.meteodata.csv}
\title{load.meteodata.csv}
\usage{
load.meteodata.csv(folder)
}
\description{
load meteodata as dataframe from folder containing CSV files
work for ZF FGEO meteostation data where single file represents one month
merge all csv files from folder in one dataframe
}
\examples{
load.meteodata.csv()
}
\keyword{meteodata}
