% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{mae}
\alias{mae}
\title{mae}
\usage{
mae(linear_model_residuals)
}
\arguments{
\item{linear_model_residuals}{linear model residuals}
}
\description{
Function that returns Mean Absolute Error
}
\examples{
mae()
}
\keyword{mae}
