% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance_modeling.R
\name{model.potential.irradiance.J}
\alias{model.potential.irradiance.J}
\title{model.potential.irradiance.J}
\usage{
model.potential.irradiance.J(plot.constants, date.variables)
}
\arguments{
\item{plot.constants}{<- c(latitude,longitude,altitude,timezone)}

\item{date.variables}{<- c(julian.date,visibility.km,humidity,temperature.K,ozone.thickness,albedo)}
}
\description{
model potential incoming irradiance in joules at given date
}
\examples{
model.potential.irradiance.J ()
}
\keyword{irradiance}
