% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{move_point_xy}
\alias{move_point_xy}
\title{move_point_xy}
\usage{
move_point_xy(x, y, azimuth, distance)
}
\arguments{
\item{x}{coord of point}

\item{y}{coord of point}

\item{azimuth}{in degrees}

\item{distance}{in map units}
}
\description{
move point xy by vector given by azimuth and distance
}
\examples{
move_point_xy()
}
\keyword{point}
