% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_operations.R
\name{mround}
\alias{mround}
\title{mround}
\usage{
mround(x, base)
}
\arguments{
\item{x}{number to round.}

\item{base}{round x to base, eg. base = 5, result rounded to the 0,5,10,15...}
}
\description{
This function round number to the given base.
}
\examples{
mround()
}
\keyword{round}
