% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_golay.R
\name{my_golay}
\alias{my_golay}
\title{my_golay}
\usage{
my_golay(model, span, p = 3)
}
\arguments{
\item{model}{a time series}

\item{span}{the number of periods to span (should be odd)}

\item{p}{the degree of the Golay polynomial}
}
\value{
the filtered time series
}
\description{
high frequency golay filtering
}
\examples{
# hc4<- my_golay(hc4.ts, 13, p=5 )
}

