% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ksmooth.R
\name{my_ksmooth}
\alias{my_ksmooth}
\title{my_ksmooth}
\usage{
my_ksmooth(timeS, span = 5, kernel = "box")
}
\arguments{
\item{timeS}{a time series}

\item{span}{the number of periods to span (should be odd)}

\item{kernel}{box or normal}
}
\value{
the filtered time series
}
\description{
high frequency ksmooth filtering
}
\examples{
# hc4<- my_ksmooth(hc4.ts, span= 13)
}

