% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lowess.R
\name{my_lowess}
\alias{my_lowess}
\title{my_lowess}
\usage{
my_lowess(timeS, span = 5)
}
\arguments{
\item{timeS}{a time series}

\item{span}{the number of periods to span}
}
\value{
the filtered time series
}
\description{
high frequency lowess filtering
}
\examples{
# hc4<- my_lowess(hc4.ts, span= 13)
}

