% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_init.R
\name{pretty_init}
\alias{pretty_init}
\title{pretty_init}
\usage{
pretty_init(pal.nr = 3)
}
\arguments{
\item{palnr}{between 1-4; one of 4 available coulour palets.}
}
\value{
NA
}
\description{
Initialize colour palet for pretty_plot routines
}
\examples{
# pretty_palet(pal.nr=4)
}

