% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_legend.R
\name{pretty_legend}
\alias{pretty_legend}
\title{pretty_legend}
\usage{
pretty_legend(kleur = c(1, 2), ...)
}
\arguments{
\item{kleur}{a vector with numbers of palet colours}

\item{...}{arguments for legend function.}
}
\value{
NA
}
\description{
put legend in upperleft corner of plot
}
\examples{
# pretty_legend(kleur=c(1,2), legend=c("a", "b"))
}

