% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_palet.R
\name{pretty_palet}
\alias{pretty_palet}
\title{pretty_palet}
\usage{
pretty_palet(FUN, kleur = 0, ...)
}
\arguments{
\item{FUN}{the fucntion to call.}

\item{kleur}{a vector with numbers of palet colours}
}
\value{
NA
}
\description{
put a palet wrapper around a standard plot function
}
\examples{
# pretty_palet(kleur=1, FUN=lines, x, y)
# pretty_abline # shorthand for pretty_palet(FUN=abline)
# pretty_text # idem
}

