% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_start.R
\name{pretty_start}
\alias{pretty_start}
\title{pretty_start}
\usage{
pretty_start(destfile, sizex, sizey, palet = 1)
}
\arguments{
\item{destfile}{output file name}

\item{sizex}{the horizontal size in px}

\item{sizey}{the vertical size in px}

\item{palet}{optional:palet to use}
}
\value{
NA
}
\description{
start writing plot output to a file
}
\examples{
# pretty_start("this.png", 500,500)
# pretty_end() # closes the file
}

