% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_operations.R
\name{randomRows}
\alias{randomRows}
\title{randomRows}
\usage{
randomRows(df, n)
}
\arguments{
\item{df}{dataframe}

\item{n}{number of rows to be returned}
}
\description{
return n random rows from dataframe
}
\examples{
randomRows()
}
\keyword{random}
\keyword{rows}
