% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_asc2tif.R
\name{raster_asc2tif}
\alias{raster_asc2tif}
\title{raster_asc2tif}
\usage{
raster_asc2tif(path)
}
\arguments{
\item{path}{directory}
}
\description{
convert all ascii rasters in folder into tif
}
\examples{
raster_asc2tif()
}
\keyword{convert}
\keyword{rasters}
