% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{raster_SJTSKtoUTM33N}
\alias{raster_SJTSKtoUTM33N}
\title{raster_SJTSKtoUTM33N}
\usage{
raster_SJTSKtoUTM33N(input.raster)
}
\arguments{
\item{raster}{raster to be projected}
}
\description{
Project raster SJTSK -> UTM33N, EPSG:  5514 -> 32633
}
\examples{
raster_SJTSKtoUTM33N()
}
\keyword{raster}
