% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{raster_UTM33NtoSJTSK}
\alias{raster_UTM33NtoSJTSK}
\title{raster_UTM33NtoSJTSK}
\usage{
raster_UTM33NtoSJTSK(input.raster)
}
\arguments{
\item{raster}{raster to be projected}
}
\description{
Project raster UTM33N -> SJTSK, EPSG: 32633 -> 5514
}
\examples{
raster_UTM33NtoSJTSK()
}
\keyword{raster}
