% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_segmentation_results.R
\name{read_first_point_xyz}
\alias{read_first_point_xyz}
\title{read_first_point_xyz}
\usage{
read_first_point_xyz(input.file)
}
\arguments{
\item{file}{name of input file}
}
\description{
Read xyz of the first point from json generated by save_first_coord.sh script
}
\examples{
read_first_point_xyz()
}
\keyword{coords}
\keyword{first}
\keyword{point}
