% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_segmentation_results.R
\name{read_rct_tree_info}
\alias{read_rct_tree_info}
\title{read_rct_tree_info}
\usage{
read_rct_tree_info(input.file)
}
\arguments{
\item{file}{name of input file}
}
\description{
Read tree info txt file produced by RayCloudTools
}
\examples{
read_rct_tree_info()
}
\keyword{input}
