% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(linear_model_residuals)
}
\arguments{
\item{linear_model_residuals}{linear model residuals}
}
\description{
Function that returns Root Mean Squared Error
}
\examples{
rmse()
}
\keyword{rmse}
