% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_manipulation.R
\name{sdf}
\alias{sdf}
\title{sdf}
\usage{
sdf(x)
}
\arguments{
\item{SpatVector}{object or any other object that can be converted by as.data.frame fci}
}
\description{
Function to view SpatVector attribute table
}
\examples{
sdf()
}
\keyword{as}
\keyword{data.frame}
\keyword{show}
