% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_cloud_functions.R
\name{set_onedrive}
\alias{set_onedrive}
\title{SET_ONEDRIVE}
\usage{
set_onedrive(folder = "none")
}
\arguments{
\item{folder}{Subfolder in OneDrive}
}
\description{
Set working directory to OneDrive folder, operating system independent
}
\examples{
set_onedrive()
}
\keyword{setwd}
