% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_cloud_functions.R
\name{setwd_icloud}
\alias{setwd_icloud}
\title{SETWD_ICLOUD}
\usage{
setwd_icloud(folder = "none")
}
\arguments{
\item{folder}{Subfolder in iCloud}
}
\description{
Set working directory to iCloud folder, operating system independent
}
\examples{
setwd_icloud()
}
\keyword{setwd}
