% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_cloud_functions.R
\name{setwd_mega}
\alias{setwd_mega}
\title{SETWD_MEGA}
\usage{
setwd_mega(folder = "none")
}
\arguments{
\item{folder}{Subfolder in MEGAsync}
}
\description{
Set working directory to MEGAsync folder, operating system independent
}
\examples{
setwd_mega()
}
\keyword{setwd}
