% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_cut.R
\name{ts_cut}
\alias{ts_cut}
\title{ts_cut}
\usage{
ts_cut(ser1, ser2)
}
\arguments{
\item{ser1}{is the series to be cut}

\item{ser2}{is the series from which the boundaries for the cut are used}
}
\value{
the truncated time series
}
\description{
cut first time series to the size of the second
}
\examples{
# co2<- ts_cut(co2.ts, rss.ts)
}

