% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_df.R
\name{ts_to_df}
\alias{ts_to_df}
\title{ts_to_df}
\usage{
ts_to_df(x)
}
\arguments{
\item{x}{a ts, zoo or xts object}
}
\value{
data frame with first column the time dimension and second column the values
}
\description{
convert a ts to a data frame for plotting
}
\examples{
# df<-ts_to_df(hc4)
# head(df, 1)
## Jaar hc4
## 1850 0.03
}

