% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_XYZ_functions.R
\name{XYZ}
\alias{XYZ}
\title{Function to perform XYZ-analysis}
\usage{
XYZ(mydata, ncol_what, ncol_by, x = 0.6, y = 0.2)
}
\arguments{
\item{mydata}{Data.frame with items sales}

\item{ncol_what}{Number of column where analysis should be performed}

\item{ncol_by}{number of column, that data should be aggregated by}

\item{x, y}{Shares coefficients of X and Y groups}
}
\value{
list with sd value for \code{ncol_what} by values from \code{ncol_by}, divided to X, Y, Z groups
}
\description{
Function to perform XYZ-analysis
}

