% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preRS.R
\name{preRS}
\alias{preRS}
\title{preRS}
\usage{
preRS(this.lm)
}
\arguments{
\item{this.lm}{a liniar model object}
}
\value{
the preRS statistic - please note this only works for a liniar model
}
\description{
calculate predictive R-squared aka leave one out Cross Validation
}
\examples{
# preRS(lm)
## 0.78
# hc4.ts <- rd.hc4()
# this.lm <- lm(hc4.ts~index(hc4.ts))
# print(summary(this.lm))
# print(preRS(this.lm))
}

